Fix Foreman legacy UI erroring out
